[feat] 도끼, 곡갱이 여부에 따라 자원 수집 기능 추가

+++ Assets/Scripts/Util/DataTable/WeaponTable.cs
@@ -21,7 +21,7 @@ public class WeaponData
     [field: SerializeField]
     public int Durability { get; set; }
     [field: SerializeField]
-    public float GatherType { get; set; }
+    public int GatherType { get; set; }
     [field: SerializeField]
     public int AttackEffect { get; set; }
     [field: SerializeField]
